handle ScancodeJsonParser
