--- core/src/test/resources/scancodefileadapter/curations_with_exclusions.yaml
@@ -1,10 +1,0 @@
-artifacts: 
-- name: "pkg/maven/com/devonfw/tools/test-project-for-deep-license-scan/0.1.0"
-  note: "some note on the curation"
-  url: "http://some/url"
-  licenses:
-    - license: MIT
-      url: https://some/license/url
-  excludedPaths:
-    - "sources/src"
-    - "foo"
